% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_data.R
\docType{data}
\name{all_site_info}
\alias{all_site_info}
\title{Useful information for every site}
\format{
A data frame with 26 rows and 12 columns:
\describe{
  \item{siteID}{A unique 3 letter code for the site}
  \item{season_seeded}{The time of year the strip was planted}
  \item{date_seeded}{The exact planting day of a site}
  \item{year_seeded}{The year the strip was planted}
  \item{hectares_in_strips}{The area of a field in prairie strips (ha)}
  \item{seeding_method}{What kind of equipment was used to seed the strip}
  \item{seed_source}{Where the seed mix was purchased}
  \item{species_seeded}{The number of species in the seed mix}
  \item{management}{Management done on a site. "Burn" indicates the site was burned}
  \item{burn_year}{The year that a site was burned}
  \item{burn_notes}{Any other relevant information about how a site is burned}
  \item{nurse_crop}{The type of crop seeded with the prairie mix, usually oats}
}
}
\usage{
all_site_info
}
\description{
A dataset containing information about every site visited for vegetation data collection
}
\keyword{datasets}
