% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{Useful information about plants identified}
\format{A data frame with 268 rows and 10 columns:
\describe{
  \item{full_name}{Scientific name - the genus and species}
  \item{common_name}{Common name of a species}
  \item{family}{Taxanomic family to which the species belongs}
  \item{native}{"Native" or "introduced" indicating origin of species}
  \item{life_cycle}{"Annual" "biennial" or "perennial" indicating the longevity of a species}
  \item{group}{Life-history group: prairie/weedy forb, C3 grass, C4 grass, sedge , rush, fern, woody}
  \item{code}{5 letter code unique to each species. First three letters of genus and first two letters of species}
  \item{alternate name}{Secondary common name}
  \item{seeds_per_oz}{Number of seeds per ounce, gathered from \href{https://www.prairiemoon.com/}{Prairie Moon Nursery}}
  \item{group_simple}{Simplified version of life history group. Consolidates C3, C4 grasses and sedges into just grasses 
  and creates "other" group for rushes, ferns, and wetland forbs}
}}
\usage{
species_list
}
\description{
A dataset containing life history and taxonomic information about plants identified in vegetation surveys.
}
\keyword{datasets}
