% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_data.R
\docType{data}
\name{strips}
\alias{strips}
\title{Useful information about strips}
\format{A data frame with 141 rows and 5 variables:
\describe{
  \item{year}{The year that the stripID was generated.}
  \item{stripID}{A numerical identifie for the strip. Each identifier begins at 1 for a given site and can be read as 
  "siteID_stripID"}
  \item{siteID}{A unique 3 letter code for the site}
  \item{area}{Area of the strip (m^2^)}
  \item{perimeter}{Perimeter of the strip (m)}
}}
\usage{
strips
}
\description{
A dataset containing the area, perimeters, and number of quadrats for each
strip within a site.
}
\keyword{datasets}
