% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_data.R
\docType{data}
\name{vegetation}
\alias{vegetation}
\title{Vegetation data for every quadrat}
\format{
A data frame with 10416 rows and 7 columns:
\describe{
  \item{year}{The year the data was collected}
  \item{quadratID}{A unique code each quadrat. Reads like "siteID_stripID_quadratNumber"}
  \item{siteID}{A unique 3 letter code for the site}
  \item{speciesID}{A unique 5 letter code for each species - consists of first 3 letters of genus, followed by 
  first two letters of species. Ex: \emph{Ratibida pinnata} = ratpi}
  \item{cover}{Visual percent cover of vegetation. Has 7 distinct catagories. }
  \item{notes}{Any notes taken during data collection}
  \item{flowering}{Whether or not a species was flowering - This was data we thought we'd collect throughout sampling
  for another project, but we really didn't end up consistently recording it. Therefore it can largely be ignored.}
}
}
\usage{
vegetation
}
\description{
A dataset containing the visual percent cover estimates for all species identified in every quadrat of every strip.
}
\keyword{datasets}
