% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_data.R
\docType{data}
\name{quadrats}
\alias{quadrats}
\title{Useful information about quadrats}
\format{A data frame with 669 rows and 5 variables:
\describe{
  \item{year}{The year the quadratID was generated. This column was added because 5 sites were visited during 
  just the second field season (2019) and some sites had to have some new quadrats generated between field seasons.}
  \item{quadratID}{A unique code for the quadrat (sampling location). Each site has 24 quadrats. 
  The number of quadrats per strip is proportional to the area of the strip. QuadratID can be read as 
  "siteID_stripNumber_quadratNumber"}
  \item{siteID}{A unique 3 letter code for the site}
  \item{easting}{Anonymized easting coordinate for quadrat location}
  \item{northing}{Anonymized northing coordate for quadrat location}
}}
\usage{
quadrats
}
\description{
A dataset containing identifiers and anonymized GPS coordinates for each quadrat visited at each site
}
\keyword{datasets}
