% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strips.R
\docType{data}
\name{strips}
\alias{strips}
\title{Strip areas, perimeters, and number of quadrats}
\format{A data frame with 120 rows and 5 variables:
\describe{
  \item{stripID}{A unique code for the strip}
  \item{siteID}{A unique code for the site}
  \item{number_quadrats}{The number of quadrats (survey locations) in that strip}
  \item{area}{Area of the strip (m^2)}
  \item{perimeter}{Perimeter of the strip (m)}
}}
\usage{
strips
}
\description{
A dataset containing the area, perimeters, and number of quadrats for each
strip within a site.
}
\keyword{datasets}
